moving helper test function to conftest.py
in order to allow arrest errors to propagate correctly they must be inside conftest.py (or import as a module.. i couldn't make in work any other way)

Testing
ran pytest 100%
